주문상세 API create

+++ src/test/java/com/example/admin/repository/ItemRepositoryTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Optional;
 
@@ -22,7 +23,7 @@ class ItemRepositoryTest extends AdminApplicationTests {
         item.setName("삼성 노트북");
         item.setTitle("삼성 노트북 A100");
         item.setContent("2019년형 노트북 입니다.");
-        item.setPrice(900000);
+        item.setPrice(BigDecimal.valueOf(900000));
         item.setBrandName("삼성");
         item.setRegisteredAt(LocalDateTime.now());
         item.setCreatedAt(LocalDateTime.now());
